Only remove a efi variable if it exists

+++ functions.sh
@@ -264,8 +264,10 @@ efivar_write() {
 
 efivar_remove() {
 	efivar_setup "${1:-}"
-	chattr -i "$var"
-	rm "$var"
+	if [ -e "$var"]; then
+		chattr -i "$var"
+		rm "$var"
+	fi
 }
 
 efivar_read() {
